requirements.txt
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu118
torchsummary==1.5.1
matplotlib==3.10.1
python3-tk